add jobs.status column
This is to allow for monitoring of jobs.

Right now, when a job is started, it is removed from the `jobs`
table. However, this will soon change to instead:

    jobs.status = 'STARTED'

and when a job is completed:

    jobs.status = 'DONE'

With the above changes, it will be possible to monitor groups of
jobs by checking that the status is 'DONE'.